+++ Import-Function.ps1
@@ -38,12 +38,12 @@ function Import-Function {
             New-Item -ItemType Directory -Path $tempDir | Out-Null
 
             # Copy function files to temp root
-            $functionFiles | Copy-Item -Destination $tempDir
+            $functionFiles | Copy-Item -Destination $tempDir -Force
 
-            # Copy moduleproject.madata.json
-            $madataFile = [System.IO.Path]::Combine($PSScriptRoot, 'moduleproject.madata.json')
-            if (Test-Path $madataFile) {
-                Copy-Item -Path $madataFile -Destination $tempDir
+            # Copy module assembler setting folder
+            $madataDir = [System.IO.Path]::Combine($PSScriptRoot, '.moduleassembler')
+            if (Test-Path $madataDir) {
+                Copy-Item -Path $madataDir -Destination $tempDir -Recurse -Force
             }
 
             # Copy resources to temp