+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.0] - 2025-12-25
+
+- Added JSON schema.
+- Moved json data for ModuleAssembler to .moduleassembler directory.
+- Refactor of all private and public functions.
+- Added Import-MAFunction.
+
 ## [0.0.1] - 2025-11-22
 
 - Initialize Fork from [ModuleTools](https://github.com/belibug/ModuleTools).